refactor: do not use separate updater class
